+++ awb_generic_sms/__manifest__.py
@@ -21,7 +21,7 @@
 
     'category': 'Sms Integration',
 
-    'version': '13.0.1.2.2',
+    'version': '13.0.1.3.0',
 
     'depends': [
         'base',
@@ -42,6 +42,7 @@
         'data/account_move_data.xml',
         'data/account_payment_data.xml',
         'data/awb_sms_template_data.xml',
+        'data/sale_subscription_data.xml',
         'sms_data_helpers/sms_sending_deactivation.xml',
     ],
     # only loaded in demonstration mode