+++ awb_subscriber_bill_automation/__init__.py
@@ -5,5 +5,5 @@
 #
 ##############################################################################
 from . import models
-from . import services
 from . import api
+from . import services