+++ awb_subscriber_bill_automation/__manifest__.py
@@ -21,7 +21,7 @@
 
     'category': 'Localization',
 
-    'version': '13.0.1.4.0',
+    'version': '13.0.1.5.0',
 
     'depends': ['crm', 'sale_management', 'sale_subscription', 'awb_subscriber_product_information'],
 
@@ -29,6 +29,7 @@
         # 'security/ir.model.access.csv',
         'data/crm_data.xml',
         'data/subscription_atm_ref_sequence.xml',
+        'data/res_users_data.xml',
         'views/crm_view.xml',
         'views/sale_view.xml',
         'views/subscription_view.xml',