+++ awb_subscriber_bill_automation/api/services/disconnection.py
@@ -4,9 +4,6 @@ from .authentication import OdooAPI
 
 import importlib
 import json
-import logging
-
-_logger = logging.getLogger(__name__)
 
 Serializer = importlib.import_module(
     "odoo.addons.odoo-rest-api"
@@ -26,6 +23,7 @@ class OdooAPI(OdooAPI):
     def _disconnect_subs(
         self, channel, discon_type, subscriptions
     ):
+
         no_permission = self._no_permission()
         if no_permission:
             return no_permission
@@ -36,11 +34,12 @@ class OdooAPI(OdooAPI):
         if invalid_params:
             return invalid_params
 
-        source = request.env[SUBSCRIPTION]._get_discon_type(
-            discon_type=discon_type.lower(), channel=channel
+        # Get the discon type
+        discon_status = request.env[SUBSCRIPTION]._get_discon_type(
+            discon_type=discon_type.lower(), channel=channel.lower()
         )
 
-        if not source:
+        if not discon_status:
             http.Response.status = "400"
             res = {
                 "errors": {
@@ -54,33 +53,19 @@ class OdooAPI(OdooAPI):
                 }
             }
 
-            _logger.info(f"----- Disconnection: <Response [{http.Response.status}]> -----")
             return json.dumps(res)
-        _logger.info(f"----- Disconnection from {source.get('name')} -----")
 
-        # Get the discon type on Odoo
-        discon_status = request.env[SUBSCRIPTION]._get_discon_type(
-            discon_type=discon_type.lower(), channel="od"
-        )
-
-        sms_ids = []
-        subscription_codes = []
-        for sub in subscriptions:
-            code = sub.get("code")
-            sms_id = sub.get("smsid")
-            if code:
-                subscription_codes.append(code)
-            elif sms_id:
-                sms_ids.append(sms_id)
+        # Extracting code in subscriptions
+        subscription_codes = [
+            sub['code'] for sub in subscriptions
+            if "code" in sub
+        ]
 
         subscription_records = request.env[SUBSCRIPTION].search([
-            "|",
-            ("code", "in", subscription_codes),
-            ("opportunity_id.jo_sms_id_username", "in", sms_ids)
+            ("code", "in", subscription_codes)
         ])
 
         if subscription_records.exists() and discon_status:
-            executed = False
             function = discon_status.get("executable")
             if function:
                 # Call function
@@ -88,19 +73,9 @@ class OdooAPI(OdooAPI):
                     "request.env[SUBSCRIPTION].%s"
                     % function
                 )
-                executed = executable(subscription_records, discon_status.get("status"))
+                executable(subscription_records, discon_status.get("status"))
 
-            if executed:
-                http.Response.status = "200"
-                status = 200
-                message = "200 OK"
-                code = 200
-            else:
-                http.Response.status = "201"
-                status = 201
-                message = "201 OK"
-                code = 201
-
+            http.Response.status = "200"
             serializer = Serializer(
                 subscription_records,
                 "{id, code, display_name, partner_id, subscription_status_subtype}",
@@ -109,9 +84,9 @@ class OdooAPI(OdooAPI):
             data = serializer.data
             res = {
                 "success": {
-                    "status": status,
-                    "message": message,
-                    "code": code,
+                    "status": 200,
+                    "message": "200 OK",
+                    "code": 200,
                     "description": "",
                     "links": {
                         "about": "",
@@ -120,24 +95,7 @@ class OdooAPI(OdooAPI):
                 "data": data,
             }
 
-            _logger.info(f"----- Disconnection: <Response [{http.Response.status}]> -----")
             return res
-
-        http.Response.status = "400"
-        res = {
-            "errors": {
-                "status": 400,
-                "code": 4002,
-                "message": "400 Bad Request",
-                "description": "Incorrect Values",
-                "links": {
-                    "about": ""
-                },
-            }
-        }
-
-        _logger.info(f"----- Disconnection: <Response [{http.Response.status}]> -----")
-        return json.dumps(res)
 
     def _check_params(self, channel, discon_type, subscriptions):
         missing_params = ""
@@ -162,5 +120,4 @@ class OdooAPI(OdooAPI):
                 }
             }
 
-            _logger.info(f"----- Disconnection: <Response [{http.Response.status}]> -----")
             return json.dumps(res)