+++ awb_subscriber_bill_automation/services/subscription_disconnection.py
@@ -29,7 +29,6 @@ class SubscriptionDisconnect(models.Model):
         database = params.get_param('odoo_database')
 
         #OAuth
-
         AUTH_URL = base_url + '/auth/'
         headers = {'Content-type': 'application/json'}