+++ streamtech_salesforce_connector/__manifest__.py
@@ -21,8 +21,8 @@
         'security/ir.model.access.csv',
         'views/crm_view.xml',
         'views/partner_view.xml',
-        'views/product_template_view.xml',
-        'views/views.xml'
+        'views/views.xml',
+        'views/product_template_view.xml'
     ],
     'installable': True,
     'application': False,