+++ streamtech_salesforce_connector/models/salesforce_importer_opportunities.py
@@ -21,7 +21,6 @@ class SalesForceImporterOpportunities(models.Model):
         sf = SalesForceConnect()
         self.sales_force = sf.connect_salesforce(self)
 
-
         # if not self.sales_force:
         #     self.connect_to_salesforce()
 
@@ -88,7 +87,6 @@ class SalesForceImporterOpportunities(models.Model):
                     Initial_Payment_Date__c,
                     Area_ODOO__c,
                     Business_Unit_Groups__c,
-                    (SELECT SLA_Activation_Actual_End_Date__c FROM opportunity.Job_Orders__r),
                     (SELECT 
                         SLA_Activation_Actual_End_Date__c, 
                         SMS_User_ID__c, SMS_Password__c 
@@ -227,9 +225,8 @@ class SalesForceImporterOpportunities(models.Model):
         if job_orders:
             for jo in job_orders.get('records', []):
                 contract_start_date = jo.get('SLA_Activation_Actual_End_Date__c', None)
-
                 sms_user_id = jo.get('SMS_User_ID__c', None)
-                sms_password = jo.get('SMS_Password__c', None)                
+                sms_password = jo.get('SMS_Password__c', None)
                 if contract_start_date and contract_term:
                     if isinstance(contract_start_date, int):
                         contract_start_date = datetime.datetime.fromtimestamp(contract_start_date/1000)
@@ -279,7 +276,7 @@ class SalesForceImporterOpportunities(models.Model):
             'company_id': zone.company_id.id,
             'team_id': sales_team_id,
             'jo_sms_id_username': sms_user_id,
-            'jo_sms_id_password': sms_password            
+            'jo_sms_id_password': sms_password
         }
 
         if contract_start_date and contract_end_date: