+++ awb_collection_reports/CHANGELOG.md
@@ -11,3 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [ADD] Computation for Current Arrears and Advances
 - [ADD] Fields for Monthly Service, Device, Security Deposit, and Others on Segmentations field
 - [ADD] Consolidate fees in matched invoices for Segmentations field
+
+## [13.0.1.0.1] - 2021-09-02
+
+- [FIX] Daily collection report fixes on payment form
+
+## [13.0.1.0.2] - 2021-09-07
+
+- [FIX] Daily collection report fixes