+++ awb_subscriber_bill_automation/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [13.0.1.5.0] - 2021-09-08
+- [IMP] SMS Provisioning and Activation flows
+- [ADD] Disconnection API
+
 ## [13.0.1.4.0] - 2021-06-10
 
 - [IMP] atm ref on subscription