+++ awb_subscriber_product_information/CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [13.0.1.3.2] - 2021-09-08
+   - [IMP] Added values for job_order_status based on what is existing in SF
+
 ## [13.0.1.3.1] - 2021-08-12
    - [ADD] Added values for subscriber_type and account_classification based on what is existing in SF
+
 ## [13.0.1.3.0] - 2021-03-25
    - Add customer number